requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
requests-socks==0.7.1
BeautifulSoup4==4.12.2
lxml==4.9.3
